TTT-2 with connections to TTT-3. Added AI gameplay, corrected format for classes, issues include undefined decisions by AI.
Added AI Gameplay, for now it is only after palyer chooses, will add function to select who goes first and have a short
wait time to give the effect that the AI is thinking. Also will add difficulty so that selections are not so random, but more precise for harder gameplay.

Classes have been reformatted so that Player and AI classes are not children of the Game class and have easier access to the gameboard grid. The reformat changes click functions and added the decision function to AI. AI function decision changes styling and symbol for AI (blue O)

ISSUES:
Randomly has UNDEFINED as AI decision. Is most likely caused by randomizing the options available to the AI. May need to use different randomizing method since decimals can be tricky.

NEEDS:
- Fix undefined issue with AI decision
- Add decision to select who goes first with flip a coin option.
- Add menu options for better user GUI (TTT-4)